Bot_telegram/requirements.txt
MouseInfo	0.1.3	
Pillow	9.4.0	
PyAutoGUI	0.9.53	
PyGetWindow	0.0.9	
PyMsgBox	1.0.9	
PyRect	0.2.0	
PyScreeze	0.1.28	
PySocks	1.7.1	
anyio	3.6.2	
async-generator	1.10	
attrs	22.2.0	
certifi	2022.12.7	
cffi	1.15.1	
chardet	5.1.0	
charset-normalizer	3.0.1	
colorama	0.4.6	
cryptography	39.0.1	
decorator	5.1.1	
filetype	1.2.0	
future	0.18.3	
h11	0.14.0	
h2	4.1.0	
hpack	4.0.0	
httpcore	0.16.3	
httpx	0.23.3	
huepy	1.2.1	
hyperframe	6.0.1	
idna	3.4	
imageio	2.25.1	
imageio-ffmpeg	0.4.8	
iniconfig	2.0.0	
mock	5.0.1	
moviepy	0.2.3.2	
numpy	1.24.2	
outcome	1.2.0	
packaging	23.0	
pip	23.0	
pluggy	1.0.0	
proglog	0.1.10	
pycparser	2.21	
pyperclip	1.8.2	
pytest	7.2.1	
python-telegram-bot	20.1	
pytweening	1.0.4	
pytz	2022.7.1	
requests	2.28.2	
requests-toolbelt	0.10.1	
responses	0.22.0	
rfc3986	2.0.0	
schedule	1.1.0	
selenium	4.8.0	
setuptools	65.5.0	
six	1.16.0	
sniffio	1.3.0	
sortedcontainers	2.4.0	
toml	0.10.2	
tornado	6.2	
tqdm	4.64.1	
trio	0.22.0	
trio-websocket	0.9.2	
types-toml	0.10.8.3	
urllib3	1.26.14	
wsproto	1.2.0	